# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Recipes API
  version: 1.0.0
  description: API para gestión de recetas, ingredientes, menús, usuarios, proveedores, inventario e impuestos.
servers:
  - url: http://localhost:4000
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
...
    User:
      type: object
      properties:
        user_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        is_active:
          type: boolean
    UserCreate:
      type: object
      required:
        - first_name
        - last_name
        - email
        - role
        - password
        - is_active
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        password:
          type: string
          format: password
        is_active:
          type: boolean
    UserUpdate:
      type: object
      required:
        - first_name
        - last_name
        - email
        - role
        - is_active
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        is_active:
          type: boolean
...
paths:
...
  /users:
    get:
      tags:
        - Usuarios
      summary: Listar usuarios
      description: |
        Recupera todos los usuarios.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      tags:
        - Usuarios
      summary: Crear usuario
      description: |
        Crea un nuevo usuario con contraseña.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener usuario por ID
      description: |
        Recupera los datos de un usuario específico.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Usuarios
      summary: Actualizar usuario
      description: |
        Modifica los datos de un usuario (sin contraseña).
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Usuarios
      summary: Eliminar usuario
      description: |
        Elimina un usuario del sistema.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...